recent popeye changes

+++ src/ibl_to_nwb/bwm_to_nwb.py
@@ -50,12 +50,11 @@ def setup_paths(one: ONE, eid: str, base_path: Path = Path.home() / "ibl_bmw_to_
         dict: _description_
     """
 
-
     subject = one.eid2ref(eid)['subject']
     paths = dict(
         output_folder = base_path / "nwbfiles" / f"sub-{subject}",
         session_folder = one.eid2path(eid),  # <- this is the folder on the main storage: /mnt/sdcepth/users/ibl/data
-        scratch_folder = base_path / 'scratch' # <- this is to be changed to /scratch on the node
+        scratch_folder = Path('/scratch') # <- this is to be changed to /scratch on the node
     )
 
     # inferred from above
@@ -110,35 +109,36 @@ def remove_uuid_from_filepath(file_path: Path) -> Path:
     else:
         return file_path
 
-def tree_copy(source_dir: Path, target_dir: Path, remove_uuid:bool=True, filter:str='.cbin'):
-    """copies all files found under source_dir (including subdirectories) to target_dir. Replicates the tree, optionally removes uuids from filenames and exludes files in filter
-    this could have include or exclude
+# def tree_copy(source_dir: Path, target_dir: Path, remove_uuid:bool=True, filter:str='.cbin'):
+#     """copies all files found under source_dir (including subdirectories) to target_dir. Replicates the tree,
+#     optionally removes uuids from filenames and exludes files in filter
+#     this could have include or exclude
 
-    Args:
-        source_dir (Path): _description_
-        target_dir (Path): _description_
-        remove_uuid (bool, optional): _description_. Defaults to True.
-        filter (str, optional): _description_. Defaults to '.cbin'.
-    """
+#     Args:
+#         source_dir (Path): _description_
+#         target_dir (Path): _description_
+#         remove_uuid (bool, optional): _description_. Defaults to True.
+#         filter (str, optional): _description_. Defaults to '.cbin'.
+#     """
 
-    for root, dirs, files in os.walk(source_dir):
-        for file in files:
-            source_file_path = Path(root) / file
-            if filter is not None:
-                if filter not in str(source_file_path):
-                    continue
+#     for root, dirs, files in os.walk(source_dir):
+#         for file in files:
+#             source_file_path = Path(root) / file
+#             if filter is not None:
+#                 if filter not in str(source_file_path):
+#                     continue
 
-            target_file_path = target_dir / source_file_path.relative_to(source_dir)
-            target_file_path.parent.mkdir(parents=True, exist_ok=True)
+#             target_file_path = target_dir / source_file_path.relative_to(source_dir)
+#             target_file_path.parent.mkdir(parents=True, exist_ok=True)
 
-            if remove_uuid:
-                parent, name = target_file_path.parent, target_file_path.name
-                name_parts = name.split(".")
-                if is_uuid_string(name_parts[-2]):
-                    name_parts.remove(name_parts[-2])
-                target_file_path = parent / ".".join(name_parts)
-            if not target_file_path.exists():
-                shutil.copy(source_file_path, target_file_path)
+#             if remove_uuid:
+#                 parent, name = target_file_path.parent, target_file_path.name
+#                 name_parts = name.split(".")
+#                 if is_uuid_string(name_parts[-2]):
+#                     name_parts.remove(name_parts[-2])
+#                 target_file_path = parent / ".".join(name_parts)
+#             if not target_file_path.exists():
+#                 shutil.copy(source_file_path, target_file_path)
 
 def filter_file_paths(file_paths: list[Path], include: list|None = None, exclude: list|None = None) ->list[Path]:
     # include filter
@@ -159,27 +159,28 @@ def filter_file_paths(file_paths: list[Path], include: list|None = None, exclude
                 [exclude_paths.append(f) for f in file_paths if excl in f.name]
         file_paths = list(set(file_paths) - set(exclude_paths))
 
-    return file_paths
+    return list(file_paths)
 
 def tree_copy(source_dir: Path, target_dir: Path, remove_uuid:bool=True, include=None, exclude=None):
     # include and exclude can be lists
-    file_paths = source_dir.rglob('*/**')
+    file_paths = list(source_dir.rglob('**/*'))
     if include is not None or exclude is not None:
         file_paths = filter_file_paths(file_paths, include, exclude)
     
     for source_file_path in file_paths:
-        target_file_path = target_dir / source_file_path.relative_to(source_dir)
-        if remove_uuid:
-            target_file_path = remove_uuid_from_filepath(target_file_path)
-        if not target_file_path.exists():
-            _logger.debug(f"copying {source_file_path} to {target_file_path}")
-            shutil.copy(source_file_path, target_file_path)
+        if source_file_path.is_file():
+            target_file_path = target_dir / source_file_path.relative_to(source_dir)
+            if remove_uuid:
+                target_file_path = remove_uuid_from_filepath(target_file_path)
+            if not target_file_path.exists():
+                _logger.debug(f"copying {source_file_path} to {target_file_path}")
+                shutil.copy(source_file_path, target_file_path)
 
 def paths_cleanup(paths: dict):
     # unlink the symlinks in the scratch folder and remove the scratch 
     # os.system(f"find {paths['session_scratch_folder']} -type l -exec unlink {{}} \;")
-    # shutil.rmtree(paths["session_scratch_folder"])
-    ...
+    _logger.debug(f"removing {paths['session_scratch_folder']}")
+    shutil.rmtree(paths["session_scratch_folder"])
 
 
 
@@ -299,8 +300,8 @@ def decompress_ephys_cbins(source_folder:Path, target_folder:Path|None=None, rem
             file_ch, = list(file_cbin.parent.glob(f'{name}*.ch'))
             
             # copies over the meta file which will still have an uuid
-            #spikeglx.Reader(file_cbin).decompress_to_scratch(file_meta=file_meta, file_ch=file_ch, scratch_dir=target_bin.parent)
-            spikeglx.Reader(file_cbin, file_meta=file_meta, file_ch=file_ch).decompress_to_scratch(scratch_dir=target_bin.parent)
+            # spikeglx.Reader(file_cbin).decompress_to_scratch(file_meta=file_meta, file_ch=file_ch, scratch_dir=target_bin.parent)
+            spikeglx.Reader(file_cbin, meta_file=file_meta, ch_file=file_ch).decompress_to_scratch(scratch_dir=target_bin.parent)
 
             if remove_uuid:
                 shutil.move(target_bin, target_bin_no_uuid)
@@ -310,7 +311,7 @@ def decompress_ephys_cbins(source_folder:Path, target_folder:Path|None=None, rem
                 file_meta_target = remove_uuid_from_filepath(target_bin.parent / file_meta.name)
                 if not file_meta_target.exists():
                     shutil.move(target_bin.parent / file_meta.name, file_meta_target)
-                    # _logger.info(f"moving {}")
+                    _logger.info(f"moving {target_bin.parent / file_meta.name} to {file_meta_target}")
 
 """
  ######   #######  ##    ## ##     ## ######## ########  ########
@@ -375,17 +376,19 @@ def convert_session(eid: str=None, one:ONE=None, revision:str=None, cleanup:bool
             subject_id = metadata["Subject"]["subject_id"]
             fname = f"sub-{subject_id}_ses-{eid}_desc-debug.nwb"
 
-    _logger.info(f"converting: {eid} with mode:{mode}")
+    _logger.info(f"converting: {eid} with mode:{mode} ... ")
     session_converter.run_conversion(
         nwbfile_path=paths["output_folder"] / fname,
         metadata=metadata,
         ibl_metadata=dict(revision=revision),
         overwrite=True,
     )
+    _logger.info(f" ... done successfully: {eid} with mode:{mode}")
 
     if cleanup:
         paths_cleanup(paths)
 
     if verify:
         check_nwbfile_for_consistency(one=one, nwbfile_path=paths["output_folder"] / fname)
+        _logger.info(f"all checks passed for {eid} with mode:{mode}")
